+++ lib/netutil/conn_pool.go
@@ -15,6 +15,12 @@ type ConnPool struct {
 	mu sync.Mutex
 	d  *TCPDialer
 
+	// concurrentDialsCh limits the number of concurrent dials the ConnPool can make.
+	// This should prevent from creating an excees number of connections during temporary
+	// spikes in workload at vmselect and vmstorage nodes.
+	// See https://github.com/VictoriaMetrics/VictoriaMetrics/issues/2552
+	concurrentDialsCh chan struct{}
+
 	name             string
 	handshakeFunc    handshake.Func
 	compressionLevel int
@@ -34,7 +40,8 @@ type connWithTimestamp struct {
 // The compression is disabled if compressionLevel <= 0.
 func NewConnPool(name, addr string, handshakeFunc handshake.Func, compressionLevel int) *ConnPool {
 	cp := &ConnPool{
-		d: NewTCPDialer(name, addr),
+		d:                 NewTCPDialer(name, addr),
+		concurrentDialsCh: make(chan struct{}, 8),
 
 		name:             name,
 		handshakeFunc:    handshakeFunc,
@@ -59,30 +66,46 @@ func (cp *ConnPool) Addr() string {
 
 // Get returns free connection from the pool.
 func (cp *ConnPool) Get() (*handshake.BufferedConn, error) {
-	var bc *handshake.BufferedConn
-	cp.mu.Lock()
-	if len(cp.conns) > 0 {
-		c := cp.conns[len(cp.conns)-1]
-		bc = c.bc
-		c.bc = nil
-		cp.conns = cp.conns[:len(cp.conns)-1]
-	}
-	cp.mu.Unlock()
-	if bc != nil {
+	if bc := cp.tryGetConn(); bc != nil {
 		return bc, nil
 	}
 
+	// Limit the number of concurrent dials.
+	// This should help https://github.com/VictoriaMetrics/VictoriaMetrics/issues/2552
+	cp.concurrentDialsCh <- struct{}{}
+	defer func() {
+		<-cp.concurrentDialsCh
+	}()
+	// Make an attempt to get already established connections from the pool.
+	// It may appear there while waiting for cp.concurrentDialsCh.
+	if bc := cp.tryGetConn(); bc != nil {
+		return bc, nil
+	}
 	// Pool is empty. Create new connection.
 	c, err := cp.d.Dial()
 	if err != nil {
 		return nil, fmt.Errorf("cannot dial %s: %w", cp.d.Addr(), err)
 	}
-	if bc, err = cp.handshakeFunc(c, cp.compressionLevel); err != nil {
+	bc, err := cp.handshakeFunc(c, cp.compressionLevel)
+	if err != nil {
 		err = fmt.Errorf("cannot perform %q handshake with server %q: %w", cp.name, cp.d.Addr(), err)
 		_ = c.Close()
 		return nil, err
 	}
 	return bc, nil
+}
+
+func (cp *ConnPool) tryGetConn() *handshake.BufferedConn {
+	cp.mu.Lock()
+	defer cp.mu.Unlock()
+	if len(cp.conns) == 0 {
+		return nil
+	}
+	c := cp.conns[len(cp.conns)-1]
+	bc := c.bc
+	c.bc = nil
+	cp.conns = cp.conns[:len(cp.conns)-1]
+	return bc
 }
 
 // Put puts bc back to the pool.
@@ -105,6 +128,8 @@ func (cp *ConnPool) Put(bc *handshake.BufferedConn) {
 
 func (cp *ConnPool) closeIdleConns() {
 	// Close connections, which were idle for more than 30 seconds.
+	// This should reduce the number of connections after sudden spikes in query rate.
+	// See https://github.com/VictoriaMetrics/VictoriaMetrics/issues/2508
 	deadline := fasttime.UnixTimestamp() - 30
 	var activeConns []connWithTimestamp
 	cp.mu.Lock()